Use Seek to read the current offset in file.
Attempting to get how many bytes the string had in the file after
we decoded it from the encoding is difficult to do correctly.

Using Seek as a ftell equivalent makes things much safer. Since we just
read from the file, the cost of the system call should be small.